+++ cichlidanalysis/analysis/PCA.py
@@ -18,6 +18,8 @@ from cichlidanalysis.analysis.run_binned_als import setup_run_binned
 from cichlidanalysis.analysis.run_feature_vector import setup_feature_vector_data
 from cichlidanalysis.io.io_feature_vector import load_diel_pattern
 from cichlidanalysis.analysis.linear_regression import run_linear_reg, plt_lin_reg
+from cichlidanalysis.plotting.plot_pca import plot_loadings, plot_2D_pc_space, plot_2D_pc_space_orig, \
+    plot_variance_explained, plot_factor_loading_matrix, pc_loadings_on_2D, plot_reconstruct_pc, plot_3D_pc_space
 # insipired by https://towardsdatascience.com/pca-using-python-scikit-learn-e653f8989e60
 
 
@@ -69,44 +71,25 @@ def fish_bin_pca_df(fish_tracks_bin, ronco_data):
     return pca_df, targets
 
 
-def plot_loadings(rootdir, pca, labels, data_input):
-    loadings = pd.DataFrame(pca.components_.T, columns=labels, index=data_input.columns)
+def fish_fv_pca_df(feature_v):
+    pca_df_fv = feature_v.drop(columns=['parental_care', 'sociality', 'distribution', 'comments/links', 'size_male',
+                                        'size_female', 'diet_contents', 'breeding', 'mouth_brooder', 'monogomous',
+                                        'habitat_details', 'genome', 'tribe', 'notes', 'species_six', 'fish_n',
+                                        'species_true', 'species_our_names'])
+    pca_df_fv, df_key = replace_cat_with_nums(pca_df_fv, col_names=['habitat', 'diet', 'cluster', 'cluster_pattern'])
+    pca_df_fv = pca_df_fv.dropna()
+    # drop movement cols
+    pca_df_fv = pca_df_fv.drop(columns=pca_df_fv.columns[pca_df_fv.columns.str.contains('move')])
+    # bouts
+    pca_df_fv = pca_df_fv.drop(columns=pca_df_fv.columns[pca_df_fv.columns.str.contains('bout')])
 
-    for pc_name in labels:
-        loadings_sorted = loadings.sort_values(by=pc_name)
-        f, ax = plt.subplots(figsize=(15, 5))
-        plt.scatter(loadings_sorted.index, loadings_sorted.loc[:, pc_name])
-        ax.set_xticklabels(loadings_sorted.index, rotation=90)
-        plt.title(pc_name)
-        ax.set_ylabel('loading')
-        sns.despine(top=True, right=True)
-        plt.axhline(0, color='gainsboro')
-        plt.tight_layout()
-        plt.savefig(os.path.join(rootdir, "loadings_{}.png".format(pc_name)), dpi=1000)
-        plt.close()
-    return loadings
+    # for species
+    pca_df_fv_sp = pca_df_fv.groupby(by='six_letter_name_Ronco').mean()
 
-
-def reconstruct_pc(data_input, pca, mu, pc_n):
-    # reconstruct the data with only pc 'n'
-    Xhat = np.dot(pca.transform(data_input)[:, pc_n - 1:pc_n], pca.components_[:1, :])
-    Xhat += mu
-    reconstructed = pd.DataFrame(data=Xhat, columns=data_input.columns)
-    f, ax = plt.subplots(figsize=(10, 5))
-    plt.plot(reconstructed)
-    plt.savefig(os.path.join(rootdir, "reconstruction_from_pc{}.png".format(pc_n)), dpi=1000)
-    plt.close()
-
-    # f, ax = plt.subplots(figsize=(10, 5))
-    # plt.plot(reconstructed.iloc[:, 0:5:])
-    # plt.savefig(os.path.join(rootdir,  "reconstruction_from_pc{}.png".format(pc_n)), dpi=1000)
-    # plt.close()
-
-    # species = 'Astbur'
-    # f, ax = plt.subplots(figsize=(10, 5))
-    # plt.plot(reconstructed.loc[:, species])
-    # plt.plot(data_input.loc[:, species])
-    # plt.close('all')
+    targets = pca_df_fv.six_letter_name_Ronco
+    all_targets = pca_df_fv.loc[:, ['six_letter_name_Ronco', 'habitat', 'diet', 'cluster', 'cluster_pattern']]
+    pca_df_fv = pca_df_fv.drop(columns=['six_letter_name_Ronco']).set_index('fish_ID')
+    return pca_df_fv, targets, all_targets, df_key, pca_df_fv_sp
 
 
 def run_pca(rootdir, data_input):
@@ -141,7 +124,7 @@ def run_pca(rootdir, data_input):
     plt.close()
 
     # plot reconstruction of pc 'n'
-    reconstruct_pc(data_input, pca, mu, 1)
+    plot_reconstruct_pc(rootdir, data_input, pca, mu, 1)
 
     # plot loadings of each pc
     loadings = plot_loadings(rootdir, pca, labels, data_input)
@@ -152,83 +135,6 @@ def run_pca(rootdir, data_input):
     # plt.close()
 
     return pca, labels, loadings, principalDf, finalDf, principalComponents
-
-
-def plot_2D_pc_space(finalDf):
-    all_species = finalDf.species.unique()
-    fig = plt.figure(figsize=(8, 8))
-    ax = fig.add_subplot(1, 1, 1)
-    ax.set_xlabel('Principal Component 1', fontsize=15)
-    ax.set_ylabel('Principal Component 2', fontsize=15)
-    ax.set_title('2 component PCA', fontsize=20)
-
-    cmap = matplotlib.cm.get_cmap('nipy_spectral')
-    for species_n, species_name in enumerate(all_species):
-        indicesToKeep = finalDf['species'] == species_name
-        ax.scatter(finalDf.loc[indicesToKeep, 'pc1'], finalDf.loc[indicesToKeep, 'pc2'],
-                   color=cmap(species_n / len(all_species)), s=50)
-    ax.legend(all_species)
-    # ax.scatter(finalDf.loc[:, 'pc1'], finalDf.loc[:, 'pc2'], s=50)
-    ax.grid()
-    plt.savefig(os.path.join(rootdir, "PCA_points_2D_space.png"), dpi=1000)
-    plt.close()
-    return
-
-
-def plot_2D_pc_space_orig(data_input):
-    # plot 2D PC space with labeled points
-    day = set(np.where(data_input.index.to_series().reset_index(drop=True) < '19:00')[0]) & set(
-        np.where(data_input.index.to_series().reset_index(drop=True) >= '07:00')[0])
-    six_thirty_am = set(np.where(data_input.index.to_series().reset_index(drop=True) == '06:30')[0])
-    seven_am = set(np.where(data_input.index.to_series().reset_index(drop=True) == '07:00')[0])
-    seven_pm = set(np.where(data_input.index.to_series().reset_index(drop=True) == '19:00')[0])
-    six_thirty_pm = set(np.where(data_input.index.to_series().reset_index(drop=True) == '18:30')[0])
-    finalDf['daynight'] = 'night'
-    finalDf.loc[day, 'daynight'] = 'day'
-    finalDf.loc[six_thirty_am, 'daynight'] = 'six_thirty_am'
-    finalDf.loc[six_thirty_pm, 'daynight'] = 'six_thirty_pm'
-    finalDf.loc[seven_am, 'daynight'] = 'seven_am'
-    finalDf.loc[seven_pm, 'daynight'] = 'seven_pm'
-
-    cmap = matplotlib.cm.get_cmap('twilight_shifted')
-    fig = plt.figure(figsize=(8, 8))
-    ax = fig.add_subplot(1, 1, 1)
-    ax.set_xlabel('Principal Component 1', fontsize=15)
-    ax.set_ylabel('Principal Component 2', fontsize=15)
-    ax.set_title('2 component PCA', fontsize=20)
-    times = finalDf.daynight
-    timepoints = times.unique()
-    # colors = ['r', 'g', 'b']
-    for time_n, time in enumerate(timepoints):
-        indicesToKeep = finalDf['daynight'] == time
-        ax.scatter(finalDf.loc[indicesToKeep, 'pc1'], finalDf.loc[indicesToKeep, 'pc2'],
-                   c=cmap(time_n / len(timepoints)), s=50)
-    ax.legend(timepoints)
-    ax.grid()
-    plt.savefig(os.path.join(rootdir, "PCA.png"), dpi=1000)
-    plt.close()
-    return
-
-
-def plot_variance_explained(principalDf, pca):
-    f, ax = plt.subplots(figsize=(5, 5))
-    cmap = matplotlib.cm.get_cmap('flare')
-    x = np.arange(0, principalDf.shape[0])
-    for col_n, col in enumerate(principalDf.columns):
-        y = principalDf.loc[:, col]
-        plt.plot(y, c=cmap(col_n / principalDf.shape[1]), label=col)
-    plt.legend()
-    plt.savefig(os.path.join(rootdir, "principalDf.png"), dpi=1000)
-    plt.close()
-
-    f, ax = plt.subplots(figsize=(5, 5))
-    plt.plot(np.cumsum(pca.explained_variance_ratio_))
-    plt.ylim([0,1])
-    ax.set_xlabel('Principal component')
-    ax.set_ylabel('Cumulative fraction of variance explained')
-    plt.savefig(os.path.join(rootdir, "explained_variance_.png"), dpi=1000)
-    plt.close()
-    return
 
 
 def replace_cat_with_nums(df, col_names):
@@ -274,13 +180,17 @@ if __name__ == '__main__':
 
     diel_patterns = load_diel_pattern(rootdir, suffix="*dp.csv")
 
-    pca_df, targets =fish_bin_pca_df(fish_tracks_bin, ronco_data)
+    pca_df, targets = fish_bin_pca_df(fish_tracks_bin, ronco_data)
+    pca_df_fv, targets, all_targets, df_key, pca_df_fv_sp = fish_fv_pca_df(feature_v)
 
-    pca, labels, loadings, principalDf, finalDf, principalComponents = run_pca(rootdir, pca_df)
+    pca, labels, loadings, principalDf, finalDf, principalComponents = run_pca(rootdir, pca_df_fv_sp)
 
-    plot_variance_explained(principalDf, pca)
-    finalDf['species'] = targets
-    plot_2D_pc_space(finalDf)
+    plot_variance_explained(rootdir, principalDf, pca)
+    finalDf['species'] = pca_df_fv_sp.loc[:, 'cluster_pattern'].reset_index(drop=True)
+    plot_2D_pc_space(rootdir, finalDf)
+    plot_3D_pc_space(rootdir, finalDf)
+    plot_factor_loading_matrix(rootdir, loadings, top_pc=3)
+    pc_loadings_on_2D(rootdir, principalComponents[:, 0:2], np.transpose(pca.components_[0:2, :]), pca_df_fv_sp.columns)
 
     loadings = loadings.reset_index().rename(columns={'index': 'features'})
     pca_df = loadings.merge(diel_patterns, on='features')
@@ -291,27 +201,3 @@ if __name__ == '__main__':
     plt_lin_reg(rootdir, pca_df.pc2, pca_df.peak, model, r_sq)
 
 
-def pc_loadings_on_2D(principalComponents, coeff, labels=None):
-    xs = principalComponents[:, 0]
-    ys = principalComponents[:, 1]
-    n = coeff.shape[0]
-    scalex = 1.0/(xs.max() - xs.min())
-    scaley = 1.0/(ys.max() - ys.min())
-    plt.scatter(xs * scalex, ys * scaley, color='y')
-    for i in range(n):
-        plt.arrow(0, 0, coeff[i, 0], coeff[i, 1], color='r', alpha=0.5)
-        if labels is None:
-            plt.text(coeff[i, 0] * 1.15, coeff[i, 1] * 1.15, "Var"+str(i+1), color='g', ha='center', va='center')
-        else:
-            plt.text(coeff[i, 0] * 1.15, coeff[i, 1] * 1.15, labels[i], color='g', ha='center', va='center')
-    plt.xlim(-1, 1)
-    plt.ylim(-1, 1)
-    plt.xlabel("PC{}".format(1))
-    plt.ylabel("PC{}".format(2))
-    plt.grid()
-    plt.savefig(os.path.join(rootdir, "pc_loadings_on_2D.png"), dpi=1000)
-    plt.close()
-    return
-
-
-pc_loadings_on_2D(principalComponents[:, 0:2], np.transpose(pca.components_[0:2, :]))
